Compute reduced header for relationships based on type info
Spark column names now only use type for properties. This allows us to
use a name-based lookup to correlate between a previous RecordHeader
and a following one, when selecting columns in the dataframe.

Expand operator now shares both the full output header as well as the
header for just the relationship parts, in order for the type filter
to work properly.