WIP Migrate to new node representation in logical and flat planner
Help writing tests by adding AsCode to get strings that can be pasted
as representations of trees and their values for simplified test
creation. This also required changing `Expr`, `IRField` and
`PropertyKeyMap` to use a customised product representation with only
one parameter list.

ScalaTest matchers were overwhelmed with the complexity of our matched
trees, which required our own `structurallyEqual` matcher.

LogicalPlannerTest still needs to be adapted to the new representation,
which is now easier because the AsCode representation of a correct
result can be pasted as the expected result.

Other remaining items:
- Update LogicalOptimizer to new rewrite rule format
- Extend and clean up the TreeNode tests