Make schema deterministically composable from individual labels
This means no label combination may have a property that does not exist
on at least one entity with only a single label from the combination.
In other words, the following is not allowed:

`(:A:B {foo: 1}), (:A), (:B)`

nor is this:

`(:A:B {foo: 1, bar: 2}), (:A {foo: 3}), (:B)`

but this is:

`(:A:B {foo: 1}), (:A {foo: 2}), (:B)`

This is a basic schema class that does not contain the full set of
property graphs, but it is still very expressive.