Remove cluster profile
Simplifies the pom by lifting module listing to the default scope and
removing the special `cluster` profile which had a different module list.
This will now build all the present modules by default.

License listing is disabled for the `spark-cypher-examples` module.
Shade plugin is moved to the modules where it is used.

Co-authored-by: Max Kießling <max.kiessling@neotechnology.com>